added a github status api hook
A set of hooks that allows bozo to submit the status of the build to github. The statuses are either `pending`, `failed` or `success`.

It can be used in conjunction with the `Jenkins` script to ensure the env variables are assigned.

The hook will only run if the `--build-server` flag is specified