Common assembly uses the build_version env variable
With the introduction of multiple version generators the CommonAssemblyInfo prepare step continued to use the version specified in the `version` file even when the version was generated with a different method.

This fixes the version hooks to run pre_prepare so the `BUILD_VERSION` environment value is available during the `CommonAssemblyInfo` step.